requirements.txt
pandas
numpy
requests
python-dotenv
pydantic
alpaca-trade-api
transformers
torch
spacy
python-telegram-bot
APScheduler
sentence-transformers
httpx